[sm-api] Return JSON file on setup
